# utils/docs/swagger.yaml
definitions:
  User.UserUpdateDetailsRequestParams:
    properties:
      birth_date:
        example: 10/01/2000
        type: string
      name:
        example: João Souza
        type: string
      phone:
        example: "71991112222"
        type: string
    required:
    - birth_date
    - name
    - phone
    type: object
  auth.LoginRequestParams:
    properties:
      email:
        example: joao@example.com
        type: string
      password:
        example: senha123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequestStep1:
    properties:
      email:
        example: joao@example.com
        type: string
      name:
        example: João Silva
        minLength: 3
        type: string
    required:
    - email
    - name
    type: object
  auth.RegisterRequestStep2:
    properties:
      confirm_password:
        example: senha123
        type: string
      hash:
        example: 72fd3951e5bd758b7619ddc62af1f052
        type: string
      password:
        example: senha123
        minLength: 6
        type: string
    required:
    - confirm_password
    - hash
    - password
    type: object
  auth.ValidationCodeRequestParams:
    properties:
      email:
        example: joao@example.com
        type: string
    required:
    - email
    type: object
  auth.VerifyCodeRequestParams:
    properties:
      email:
        example: joao@example.com
        type: string
    required:
    - email
    type: object
  controllers.TokenResponse:
    properties:
      token:
        type: string
    type: object
  models.User:
    properties:
      birth_date:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  resources.UserResource:
    properties:
      birth_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Realiza login de usuário
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequestParams'
      produces:
      - application/json
      responses:
        "201":
          description: Resposta de login
          schema:
            $ref: '#/definitions/controllers.TokenResponse'
      summary: Logar usuário
      tags:
      - Autenticação
  /api/auth/logout:
    get:
      consumes:
      - application/json
      description: Desloga o usuário atual
      produces:
      - application/json
      responses:
        "200":
          description: Usuário deslogado com sucesso
          schema:
            type: object
      summary: Deslogar usuário
      tags:
      - Autenticação
  /api/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Retorna um novo token de acesso válido por 1 hora
      parameters:
      - description: Token de refresh
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Novo token de acesso
          schema:
            type: string
      summary: Refresca o token de acesso
      tags:
      - Autenticação
  /api/auth/register_step1:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário sem senha
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequestStep1'
      produces:
      - application/json
      responses:
        "201":
          description: Usuário registrado
          schema:
            $ref: '#/definitions/resources.UserResource'
      summary: Registrar usuário (Step 1)
      tags:
      - Autenticação
  /api/auth/register_step2:
    post:
      consumes:
      - application/json
      description: Define a senha do usuário
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequestStep2'
      produces:
      - application/json
      responses:
        "200":
          description: Senha definida
          schema:
            $ref: '#/definitions/resources.UserResource'
      summary: Registrar usuário (Step 2)
      tags:
      - Autenticação
  /api/auth/send-validation-code:
    post:
      description: Envia código de validação para o e-mail fornecido.
      parameters:
      - description: e-mail
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ValidationCodeRequestParams'
      responses:
        "200":
          description: 'Código enviado" example="{\"message\": \"Código enviado com
            sucesso\"}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Envia código de validação
      tags:
      - Autenticação
  /api/auth/verify-validation-code:
    post:
      description: Verifica se o código de validação é válido.
      parameters:
      - description: E-mail do usuário
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: email, code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.VerifyCodeRequestParams'
      responses:
        "200":
          description: Código validado com sucesso
          schema:
            type: object
      summary: Verifica código de validação
      tags:
      - Autenticação
  /api/users:
    get:
      description: Obtém uma lista de usuários cadastrados
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Listar usuários
      tags:
      - Users
  /api/users/{id}:
    get:
      description: Obtém informações de um usuário em específico
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Listar usuário
      tags:
      - Users
    put:
      description: Atualiza informações de um usuário em específico
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/User.UserUpdateDetailsRequestParams'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Atualizar usuário
      tags:
      - Users
swagger: "2.0"